requirements.txt
python-dotenv
streamlit

langchain-core
langchain-community
langchain-text-splitters
langchain-nvidia-ai-endpoints

faiss-cpu
pypdf
